+++ lib/exceptions/handler/handler.dart
@@ -2,10 +2,13 @@ import 'dart:async';
 
 import '../../config/logger.dart';
 import '../exceptions.dart';
+import '../git_exceptions.dart';
 
 Future<void> handle(FutureOr<void> Function() function) async {
   try {
     await function();
+  } on GitException catch (e) {
+    log.e("Problems related to git found\n$e");
   } on AppException catch (e) {
     log.e(e);
   } catch (e, s) {